Mediator: Restructured. Added ReadMe.md file
