+++ src/Mediator/mediator/Mediator.java
@@ -1,4 +1,6 @@
-package Mediator;
+package Mediator.mediator;
+
+import Mediator.colleague.Colleague;
 
 /**
  * Created by navid on 12/14/17.
